+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
   - postgres: simple CRUD support for LargeObject API
   - kotlin-sqlobject: fix package declaration of RegisterKotlinMappers
   - LocalTransactionHandler: bind more closely to Handle for performance and to avoid leaks
+  - ResultProducers: give access to produced result on context
 
 # 3.13.0
   - Kotlin: respect default values in methods when passed null, #1690